+++ Sources/Example_10/main.swift
@@ -35,7 +35,7 @@ public class Example_10 {
         let page = Page(pdf, Letter.PORTRAIT)
 
         image1.setLocation(90.0, 35.0)
-        image1.scaleBy(0.75)
+        image1.scaleBy(1.25)        // The image is 120 DPI, so 0.6 of its pixels
         image1.drawOn(page)
 
         // To rotate the column, add it to a Container and rotate that; see Example_35.